Add excplicit cast from float to uint8_t for color

+++ ouzel/Node.cpp
@@ -157,7 +157,7 @@ namespace ouzel
             {
                 if (currentLayer->getCamera())
                 {
-                    graphics::Color drawColor(color.r, color.g, color.b, color.a * opacity);
+                    graphics::Color drawColor(color.r, color.g, color.b, static_cast<uint8_t>(color.a * opacity));
 
                     for (const DrawablePtr& drawable : drawables)
                     {
